Remove unused screens and components; update navigation for project editing and not found routes
